Variable-sized reg frames 30 - fix initial context creation
* the initial context was created w/o a register store
* this caused some memory corruption in src tests, like t/src/extend_13 (ppc error)
* reuse normal allocation code, allocate 32 regs each for first frame


git-svn-id: https://svn.parrot.org/parrot/trunk@9771 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ t/src/extend.t
@@ -535,7 +535,7 @@ caught
 back
 OUTPUT
 
-unlink "$temp.pasm", "$temp.pbc";
+unlink "$temp.pasm", "$temp.pbc" unless $ENV{POSTMORTEM};
 
 
 1;
